requirements.txt
# Core Python Data stack
pandas>=2.0.0
numpy>=1.24.0

# Database
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0

# Data Fetching
yfinance>=0.2.0

# Web Dashboard / API
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.30.0
